+++ src/main/java/uk/gov/ons/collection/controller/FormDefinitionController.java
@@ -12,6 +12,7 @@ import uk.gov.ons.collection.entity.ContributorEntity;
 import uk.gov.ons.collection.entity.FormDefinitionEntity;
 import uk.gov.ons.collection.exception.DataNotFoundException;
 import uk.gov.ons.collection.service.FormDefinitionService;
+import uk.gov.ons.collection.utilities.UrlParameterBuilder;
 
 import java.util.*;
