+++ src/main/java/uk/gov/ons/collection/entity/PeriodOffsetQuery.java
@@ -3,7 +3,6 @@ package uk.gov.ons.collection.entity;
 import java.util.HashMap;
 import java.util.Map;
 
-import uk.gov.ons.collection.controller.qlQueryBuilder;
 
 
 public class PeriodOffsetQuery {
